+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 + Change Jieba::Locate to be static function.
 + Change the return value of KeywordExtractor::Extract from bool to void.
++ Add KeywordExtractor::Word and add more overrided KeywordExtractor::Extract
 
 ## v4.5.3